+++ shipping-service/src/main/java/edu/miu/cs590/shippingservice/ShipmentService.java
@@ -3,10 +3,8 @@ package edu.miu.cs590.shippingservice;
 
 public interface ShipmentService {
 
-
-
      void processShipment(ShipmentRequest shipmentRequest);
-     String checkShipmentStatus(int  orderId);
+     String checkShipmentStatus(Long  orderId);
